requirements.txt
Django==1.9.6
PyYAML==3.11
argparse==1.2.1
django-rest-swagger==0.3.7
djangorestframework==3.3.3
six==1.10.0
wsgiref==0.1.2
